fsck.gfs2: Clean up metawalk_check_data and undo_check_data argument passing
Move the metatype check out of hdr_size and into a new function (should_check).
Pass the block's height to metawalk_check_data and compute the rest inside this
function.  Likewise, pass the bh and height to undo_check_data and compute the
rest inside this function.  With that, check_metatree doesn't need to know
about hdr_size anymore.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>