fsck.gfs2: More flexible metalist handling in check_metatree
In check_metatree, allocate the metalist on the stack and only initialize it as
far as needed.  Adjust free_metalist accordingly.  This doesn't change much at
the moment, but it will help with the upcoming extent-based inode support
because the extent tree isn't limited to a height of GFS2_MAX_META_HEIGHT (and
is expected to be much smaller than that in practice).

Get rid of a duplicate but harmless free_metalist call in
build_and_check_metalist.

Get rid of some unnecessary &metalist[0] obfuscation.  (There are several more
instances of &array[0] and similar that I didn't change.)

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>