libgfs2: Remove i_di from struct gfs2_inode
To stop storing native-endian data in a struct gfs2_dinode, its fields
are duplicated into struct gfs2_inode using uint*_t types and the
in/out/print functions updated to work directly with the struct
gfs2_inode. This requires a few intermediate gfs2_inums to be used in
places that previously used i_di.di_num as a whole. They will be
resolved once a native-endian replacement for gfs2_inums is added.

Unfortunately the large size of this patch is necessary as the switch is
an all-or-nothing deal, but most of it is fairly mechanical.

Signed-off-by: Andrew Price <anprice@redhat.com>