libgfs2: Remove big-endian structs from struct rgrp_tree
Replace them with native-endian fields that combine the struct
gfs2_rindex, struct gfs2_rgrp and struct gfs_rgrp. Some fields (e.g.
rg_crc) are omitted because they are unneeded other than when the struct
is read and written. This reduces the memory usage of the resource group
tree in the utils. Convert the relevant gfs2_(struct)_in/out() functions
to use the new fields.

Signed-off-by: Andrew Price <anprice@redhat.com>