Make sure i_bh is set after lgfs2_gfs_inode_get()
This function accepts a char * instead of a bh and users expect i_bh to
be set as it is in lgfs2_inode_get(). Fixes a segfault when checking
gfs1 filesystems. This is kind of a temporary measure as i_bh should be
removed in future in favour of separately managed i/o.

Signed-off-by: Andrew Price <anprice@redhat.com>