libgfs2: Move struct printing functions out of libgfs2
Now that libgfs2 itself doesn't use these functions, they can be moved
into the two utils that use them. This does duplicate the code but that
means the utils are free to change the way that they're printed to suit
their interfaces. We're still using the "print_it" extern that libgfs2.h
requires but that can be removed next.

Signed-off-by: Andrew Price <anprice@redhat.com>