tests: Add an option to clean up the testvol file unconditionally
Failed tests skip cleanup, so if the final test doesn't remove the
testvol then it is left around in the test environment. For CI we want
to remove it so provide a --always-clean-testvol option which cleans up
unconditionally in an EXIT trap routine.

This can be used via 'make check TOPTS=--always-clean-testvol'

Signed-off-by: Andrew Price <anprice@redhat.com>