Add rpm make rule for testing
'make rpm' generates a minimal gfs2-utils.spec from gfs2-utils.spec.in
and builds the rpm directly from the source tree.

The spec files are kept in tests/ to make sure they're not used for
anything else. The rpm rule depends on the clean rule as the spec file
needs to be regenerated each time in order to set the correct git
version.

An example resulting rpm is:

  rpmbuild/x86_64/gfs2-utils-3.5.1.1.dev-2.ge3f9cb1.x86_64.rpm

Where 3.5.1.1.dev is the package version and 2.ge3f9cb1 is generated by
'git describe', where HEAD is 2 commits along from the latest annotated
tag and its commit hash is e3f9cb1. If there are outstanding changes in
the working tree, ".dirty" is appended after the commit hash.

When the source tree is not a git tree, the appended rpm release is 0.

Signed-off-by: Andrew Price <anprice@redhat.com>