fsck.gfs2: Fix xattr offset checks in p1_check_eattr_entries
Valid ea_header offsets fall within the block, at the block boundary,
but not in the final 15 bytes of the block as that would result in a
partial ea_header. Make sure these cases are all taken into account in
the ea_rec_len checks in p1_check_eattr_entries(). Also improve logging
of erroneous values.

Signed-off-by: Andrew Price <anprice@redhat.com>