libgfs2: Enable use in C++ code
'new' is a keyword in C++ and 'register' was removed from it, so don't
use those in libgfs2.h and osi_tree.h. Ideally we would hide the tree
implementation behind an abstraction but the struct embedding in
libgfs2.h means that libgfs2 users need to know about the
implementation. This allows C++ code to statically link with libgfs2
(it's far from ready for shipping as a dynamic library).

Signed-off-by: Andrew Price <anprice@redhat.com>