make rpm: Don't install to /usr/sbin unconditionally
Fedora's merge of /usr/sbin into /usr/bin breaks 'make rpm' as it now
expects everything to be installed to /usr/bin at the end but we're
installing into /usr/sbin. Switch to using the configure macro which
sets all of the path variables for us, and update some hardcoded paths
in the spec file.

Signed-off-by: Andrew Price <anprice@redhat.com>