gfs2_edit: Allow setting flags fields using symbols
Parse combinations of symbolic flags to set flags fields.

The separator is chosen to be '|' as it's consistent with C. It has the
drawback of needing to be escaped in command lines but unfortunately the
argument parsing code treats anything with '/' as a path so the
separator used in output cannot be used. Adding a third separator would
be even more confusing.

Signed-off-by: Andrew Price <anprice@redhat.com>